Expose JS string literal as a String extension
Foundation helpers should be ergonomic extensions on existing types, not
bare top-level functions. Replace the public func cmuxJavaScriptStringLiteral
with String.javaScriptStringLiteral and update the call site. Behavior is
unchanged.
